Called recoverhp in healables classes

+++ Assets/Scripts/Consumables/Healables/IHealable.cs
@@ -3,6 +3,6 @@ using System.Collections;
 
 namespace Consumables.Healables {
 	public interface IHealable : IConsumable {
-		int HealthPercentage { get; }
+		float HealthPercentage { get; }
 	}
 }
